UI DetailedController Update
