+++ TheDogApp/Controller/HomeController.swift
@@ -16,6 +16,8 @@ class HomeController: UIViewController {
         return UserDefaults.standard.array(forKey: "imageKey")as? [String] ?? ["primeiraTela"]
     }
     
+    private let cell = "MyCollectionViewCell"
+    
     @IBOutlet weak var collectionView: UICollectionView!
     @IBOutlet weak var spinner: UIActivityIndicatorView!
 
@@ -43,10 +45,10 @@ class HomeController: UIViewController {
     }
     
     private func collectionUpdate(){
-        collectionView.register(CollectionViewCell.nib(), forCellWithReuseIdentifier: CollectionViewCell.identifier)
+        collectionView.register(CollectionViewCell.nib(), forCellWithReuseIdentifier: cell)
         collectionView.delegate = self
         collectionView.dataSource = self
-        collectionView.reloadData()
+//        collectionView.reloadData()
     }
     
     override func viewDidLoad() {
@@ -73,7 +75,8 @@ extension HomeController: UICollectionViewDataSource{
     }
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: CollectionViewCell.identifier, for: indexPath) as! CollectionViewCell
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cell, for: indexPath) as? CollectionViewCell else {return UICollectionViewCell()}
+        
         DispatchQueue.main.async() {
             cell.labelBreed.text = self.getName()[indexPath.row]
             cell.imageView.downloaded(from: self.getImage()[indexPath.row])
@@ -82,24 +85,5 @@ extension HomeController: UICollectionViewDataSource{
     }
 }
 
-extension UIImageView {
-    func downloaded(from url: URL, contentMode mode: ContentMode = .scaleAspectFit) {
-        contentMode = mode
-        URLSession.shared.dataTask(with: url) { data, response, error in
-            guard
-                let httpURLResponse = response as? HTTPURLResponse, httpURLResponse.statusCode == 200,
-                let mimeType = response?.mimeType, mimeType.hasPrefix("image"),
-                let data = data, error == nil,
-                let image = UIImage(data: data)
-                else { return }
-            DispatchQueue.main.async() { [weak self] in
-                self?.image = image
-            }
-        }.resume()
-    }
-    func downloaded(from link: String, contentMode mode: ContentMode = .scaleAspectFit) {
-        guard let url = URL(string: link) else { return }
-        downloaded(from: url, contentMode: mode)
-    }
-}
+
